requirements.txt
# Core Django packages
Django==4.2.7
djangorestframework==3.14.0

# Database and ORM
psycopg2-binary==2.9.7  # For PostgreSQL (production ready)

# CORS handling for React frontend
django-cors-headers==4.3.1

# File handling and image processing
Pillow==10.0.1

# Development and debugging tools
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# API documentation
drf-spectacular==0.26.5

# Authentication and permissions
djangorestframework-simplejwt==5.3.0

# Environment variables management
python-decouple==3.8

# Testing
pytest==7.4.3
pytest-django==4.5.2
factory-boy==3.3.0

# Code quality
flake8==6.1.0
black==23.9.1

# File uploads and storage
django-storages==1.14.2  # For cloud storage if needed later

# ORCID integration (will be added when implementing research app)
# requests==2.31.0  # For API calls to ORCID

# Payment processing (will be added when implementing training/services)
# stripe==6.7.0  # For payment integration

# Additional utilities
django-filter==23.3  # For API filtering
django-crispy-forms==2.1  # For better forms in admin

# Internationalization and Translation
django-modeltranslation==0.19.15
django-rosetta==0.10.2

# Database URL parsing
dj-database-url==2.1.0
